Moved utility functions for testing.
